media:i2c: Fix more coverity issues
Fix bit shift
Remove unused code

Coverity ID 36335
Coverity ID 36346
Coverity ID 32212
Coverity ID 1706467
Coverity ID 1706468

Bug 200192387

Reviewed-on: http://git-master/r/1317485
GVS: Gerrit_Virtual_Submit

+++ drivers/media/i2c/ov9281.c
@@ -474,10 +474,10 @@ fail:
 static int ov9281_set_coarse_time(struct ov9281 *priv, s32 val)
 {
 	ov9281_reg regs[4];
-	u16 coarse_time;
+	u32 coarse_time;
 	int err;
 
-	coarse_time = (u16)val;
+	coarse_time = (u32)val;
 
 	dev_dbg(&priv->i2c_client->dev,
 		 "%s: coarse_time: %d\n", __func__, coarse_time);
